(GH-799) Revive docfx scripts
Previously the document generation scripts for docfx were broken e.g. Unable to
download docfx from github due to TLS.  This commit;
* Updates the BuildDocs script to properly clean and download docfx from Github
* Updates the docfx.json file to use the correct naming (cwd -> src) and to
  use the correct project files
* Updates gitignore to ignore docfx working directory
* Adds two build tasks (BuildDocs and ServeDocs) to build and serve
  autogenerated documentation
* Adds additional configuration switch for Invoke-Build for legacy applications,
  like docfx, which use MSBuild as a library but it needs the dotNet Core
  versions of MSBuild instead of VS 2015/2017

+++ scripts/BuildDocs.ps1
@@ -11,18 +11,46 @@ $docsRepoPath = [System.IO.Path]::GetFullPath("$PSScriptRoot\..\docs\_repo")
 # Ensure the tools path exists
 mkdir $toolsPath -Force | Out-Null
 
-if (![System.IO.File]::Exists($docfxExePath)) {
-    # Download DocFX
-    Remove-Item -Path "$docfxBinPath" -Force -ErrorAction Stop | Out-Null
-    (new-object net.webclient).DownloadFile('https://github.com/dotnet/docfx/releases/download/v1.8/docfx.zip', "$docfxZipPath")
+if (!(Test-Path -Path $docfxBinPath)) {
+    mkdir $docfxBinPath -Force | Out-Null
+}
 
-    # Extract the archive
-    Expand-Archive $docfxZipPath -DestinationPath $docfxBinPath -Force -ErrorAction "Stop"
+if (!(Test-Path -Path $docfxExePath)) {
+    # Download DocFX
+    if (Test-Path -Path $docfxBinPath) { Remove-Item -Path "$docfxBinPath" -Force -ErrorAction Stop | Out-Null }
+    # Github uses TLS 1.2
+    $originalSecurityProtocol = [Net.ServicePointManager]::SecurityProtocol
+    try {
+        [Net.ServicePointManager]::SecurityProtocol = [Net.ServicePointManager]::SecurityProtocol -bor [Net.SecurityProtocolType]::Tls12
+        Write-Verbose "Downloading docfx..."
+        # The docFX 1.x release does generate documentation but is incomplete as it doesn't understand dotNet 2.1 (https://github.com/dotnet/docfx/releases/download/v1.9.4/docfx.zip)
+        # The docFX 2.x release errors during generation as it doesn't understand dotNet 2.1 (https://github.com/dotnet/docfx/releases/download/v2.40.3/docfx.zip)
+        # 2.1 support seems to be slated for docfx 3.x release - https://github.com/dotnet/docfx/projects/1
+        #
+        # For now use the 1.x series of docfx
+        Invoke-WebRequest -Uri 'https://github.com/dotnet/docfx/releases/download/v1.9.4/docfx.zip' -OutFile $docfxZipPath -ErrorAction Stop -UseBasicParsing
+
+        # Extract the archive
+        Expand-Archive $docfxZipPath -DestinationPath $docfxBinPath -Force -ErrorAction "Stop"
+    }
+    Finally {
+        [Net.ServicePointManager]::SecurityProtocol = $originalSecurityProtocol
+    }
 }
 
 # Clean the _site folder if necessary
 if ($Clean.IsPresent) {
-    Remove-Item -Path $sitePath -Force -Recurse | Out-Null
+    if (Test-Path -Path $sitePath) { Remove-Item -Path $sitePath -Force -Recurse | Out-Null }
+
+    # Clean docfx object caches
+    @("src\PowerShellEditorServices",
+      "src\PowerShellEditorServices.Channel.WebSocket",
+      "src\PowerShellEditorServices.Host",
+      "src\PowerShellEditorServices.Protocol",
+      "src\PowerShellEditorServices.VSCode") | ForEach-Object -Process {
+        $docCache = [System.IO.Path]::GetFullPath("$PSScriptRoot\..\" + $_ + "\obj\xdoc")
+        if ([System.IO.Directory]::Exists($docCache)) { Remove-Item -Force -Recurse -Path $docCache | Out-Null }
+    }
 }
 
 # Build the metadata for the C# API
